Trabajo Práctico listo para entregar

+++ disparo.c
@@ -9,14 +9,14 @@
 
 struct disparo{
     figura_t *fig;
-    double pos[2];
-    double vel[2];
+    float pos[2];
+    float vel[2];
     double angulo;
     double tiempo;
     bool friendly;
 };
 
-disparo_t *disparo_crear(double pos_x, double pos_y, double vel_x, double vel_y, double angulo, bool friendly){
+disparo_t *disparo_crear(float pos_x, float pos_y, float vel_x, float vel_y, double angulo, bool friendly){
     disparo_t *disparo = malloc(sizeof(disparo_t)*1);
     if(disparo == NULL) return NULL;
 
@@ -66,11 +66,11 @@ double disparo_get_tiempo(disparo_t *disparo){
     return disparo->tiempo;
 }
 
-double disparo_get_pos_x(disparo_t *disparo){
+float disparo_get_pos_x(disparo_t *disparo){
     return disparo->pos[X];
 }
 
-double disparo_get_pos_y(disparo_t *disparo){
+float disparo_get_pos_y(disparo_t *disparo){
     return disparo->pos[Y];
 }
 
@@ -78,7 +78,7 @@ bool disparo_friendly(disparo_t *disparo){
     return disparo->friendly;
 }
 
-double distancia_a_disparo(disparo_t *disparo, double pos_x, double pos_y){
+float distancia_a_disparo(disparo_t *disparo, float pos_x, float pos_y){
     return distancia_entre_puntos(pos_x, pos_x, disparo->pos[X], disparo->pos[Y]);
 }
